domparsing/serializing_xhtml_doctypevariation4.xhtml

<!DOCTYPE  MyName  [   <!ELEMENT  e ( #PCDATA)>   ]  >
<html xmlns="http://www.w3.org/1999/xhtml">
 <head>
  <title>DOM Serialization in XHTML - variation 4</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <div id="log"></div>
  <script><![CDATA[
  test(function(){
    // Note: line endings are not interoperable (for windows-systems, some
    // browsers let literal \r\n through their XML parser, while others
    // convert \r\n to \n.

    // Set testing expectations based on platform capability
    var expectedSerialization = (typeof document.doctype.internalSubset != undefined) ?
        "<!DOCTYPE MyName [   <!ELEMENT  e ( #PCDATA)>   ]>" :
        "<!DOCTYPE MyName>";
    assert_equals(new XMLSerializer().serializeToString(document.doctype), expectedSerialization);
  }, "Parsed DOCTYPE serialized correctly");

  // ]]>
  </script>
 </body>
</html>
